requirements.txt
fake-bpy-module==20260423
